working version

+++ src/main/java/hu/elte/finances/repositories/TransactionRepository.java
@@ -9,6 +9,6 @@ import hu.elte.finances.entities.Transaction;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TransactionRepository extends CrudRepository<Transaction, Long>, JpaSpecificationExecutor<Transaction> {
+public interface TransactionRepository extends CrudRepository<Transaction, Integer>, JpaSpecificationExecutor<Transaction> {
     
 }
